Added robust mechanism to AccessKey (#19)
* Added robust mechanism to AccessKey .

* Fix

---------

+++ upsonic_on_prem/utils/accesskey.py
@@ -12,16 +12,33 @@ class AccessKey:
     def __init__(self, key):
         self.key = key
 
+    def set_robust(self, robust):
+        return self._set(self.key + ":robust", robust)
+
+    @property
+    def robust(self):
+        return self._get(self.key + ":robust") == True
+
+
     def _set(self, key, value):
+        if self.robust:
+            return False
+
         return storage.set(key, value)
 
     def _get(self, key):
         return storage.get(key)
 
     def _delete(self, key):
+        if self.robust:
+            return False
+
         return storage.delete(key)
 
     def _keys(self):
+        if self.robust:
+            return False
+
         return storage.keys()
 
 
@@ -125,4 +142,6 @@ class AccessKey:
         return control
 
 if admin_key is not None:
-    AccessKey(admin_key).set_is_admin(True)
+    the_admin = AccessKey(admin_key)
+    the_admin.set_is_admin(True)
+    the_admin.set_robust(True)
